+++ clang-tidy/nodecpp/NakedPtrFromFunctionCheck.h
@@ -1,4 +1,5 @@
-//===--- NakedPtrFromFunctionCheck.h - clang-tidy------------------------*- C++ -*-===//
+//===--- NakedPtrFromFunctionCheck.h - clang-tidy------------------------*- C++
+//-*-===//
 //
 //                     The LLVM Compiler Infrastructure
 //